+++ display.h
@@ -1,19 +1,7 @@
 #ifndef DISPLAY_H
 #define DISPLAY_H
 
-#include <gtk/gtk.h>
-#include <gdk/gdk.h>
 #include <cairo.h>
-
-extern GtkWidget *main_window;
-
-extern void import_dialog(GtkWidget *, gpointer);
-extern void report_error(GError* error);
-
-extern GtkWidget *dive_profile_widget(void);
-extern GtkWidget *dive_info_frame(void);
-extern GtkWidget *extended_dive_info_widget(void);
-extern GtkWidget *equipment_widget(void);
 
 extern void repaint_dive(void);
 extern void do_print(void);
@@ -35,5 +23,6 @@ struct graphics_context {
 };
 
 extern void plot(struct graphics_context *gc, int w, int h, struct dive *dive);
+extern void set_source_rgb(struct graphics_context *gc, double r, double g, double b);
 
 #endif
